Move ReferenceNames() to the resolver.go
