+++ resolver/resolver_test.go
@@ -4,7 +4,10 @@ import (
 	"path/filepath"
 	"testing"
 
+	exprv1 "google.golang.org/genproto/googleapis/api/expr/v1alpha1"
+
 	"github.com/google/go-cmp/cmp"
+	exprv1 "google.golang.org/genproto/googleapis/api/expr/v1alpha1"
 	"google.golang.org/genproto/googleapis/rpc/code"
 
 	"github.com/mercari/grpc-federation/internal/testutil"
@@ -3125,4 +3128,105 @@ func getContentProtoBuilder(t *testing.T) *testutil.FileBuilder {
 				Build(t),
 		)
 	return pb
+}
+
+func TestValidationError_ReferenceNames(t *testing.T) {
+	v := &resolver.ValidationError{
+		Rule: &resolver.CELValue{
+			CheckedExpr: &exprv1.CheckedExpr{
+				ReferenceMap: map[int64]*exprv1.Reference{
+					0: {Name: "name1"},
+				},
+			},
+		},
+		Details: []*resolver.ValidationErrorDetail{
+			{
+				Rule: &resolver.CELValue{
+					CheckedExpr: &exprv1.CheckedExpr{
+						ReferenceMap: map[int64]*exprv1.Reference{
+							0: {Name: "name2"},
+						},
+					},
+				},
+				PreconditionFailures: []*resolver.PreconditionFailure{
+					{
+						Violations: []*resolver.PreconditionFailureViolation{
+							{
+								Type: &resolver.CELValue{
+									CheckedExpr: &exprv1.CheckedExpr{
+										ReferenceMap: map[int64]*exprv1.Reference{
+											0: {Name: "name3"},
+										},
+									},
+								},
+								Subject: &resolver.CELValue{
+									CheckedExpr: &exprv1.CheckedExpr{
+										ReferenceMap: map[int64]*exprv1.Reference{
+											0: {Name: "name4"},
+										},
+									},
+								},
+								Description: &resolver.CELValue{
+									CheckedExpr: &exprv1.CheckedExpr{
+										ReferenceMap: map[int64]*exprv1.Reference{
+											0: {Name: "name5"},
+										},
+									},
+								},
+							},
+							{
+								Type: &resolver.CELValue{
+									CheckedExpr: &exprv1.CheckedExpr{
+										ReferenceMap: map[int64]*exprv1.Reference{
+											0: {Name: "name3"},
+										},
+									},
+								},
+								Subject: &resolver.CELValue{
+									CheckedExpr: &exprv1.CheckedExpr{
+										ReferenceMap: map[int64]*exprv1.Reference{
+											0: {Name: "name4"},
+										},
+									},
+								},
+								Description: &resolver.CELValue{
+									CheckedExpr: &exprv1.CheckedExpr{
+										ReferenceMap: map[int64]*exprv1.Reference{
+											0: {Name: "name6"},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	expected := []string{
+		"name1",
+		"name2",
+		"name3",
+		"name4",
+		"name5",
+		"name6",
+	}
+	got := v.ReferenceNames()
+	// the map order is not guaranteed in Go
+	if len(expected) != len(got) {
+		t.Errorf("the number of reference names is different: expected: %d, got: %d", len(expected), len(got))
+	}
+	for _, e := range expected {
+		var found bool
+		for _, g := range got {
+			if g == e {
+				found = true
+				break
+			}
+		}
+		if !found {
+			t.Errorf("%q does not exist in the received reference names: %v", e, got)
+		}
+	}
 }
